MyFunctionProj/requirements.txt
azure-functions
azure-storage-blob
python-magic
